+++ resources/views/web/auth/index.blade.php
@@ -1,0 +1,5 @@
+@extends('web.app')
+@section('title', 'Página Inicial')
+@section('content')
+    @livewire('web.auth')
+@endsection